Move db into nyaa.extensions
* Switch to `db.init_app(app)`
* Change all `from nyaa import db` imports to `from nyaa.extensions import db`
