+++ nyaa/__init__.py
@@ -4,16 +4,12 @@ import os
 import flask
 from flask_assets import Bundle, Environment  # noqa F401
 from flask_debugtoolbar import DebugToolbarExtension
-from flask_sqlalchemy import SQLAlchemy
 
 from nyaa import fix_paginate  # noqa F401
+from nyaa.extensions import db
 
 app = flask.Flask(__name__)
 app.config.from_object('config')
-
-# Database
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['MYSQL_DATABASE_CHARSET'] = 'utf8mb4'
 
 # Don't refresh cookie each request
 app.config['SESSION_REFRESH_EACH_REQUEST'] = False
@@ -63,7 +59,10 @@ app.jinja_env.add_extension('jinja2.ext.do')
 app.jinja_env.lstrip_blocks = True
 app.jinja_env.trim_blocks = True
 
-db = SQLAlchemy(app)
+# Database
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+app.config['MYSQL_DATABASE_CHARSET'] = 'utf8mb4'
+db.init_app(app)
 
 assets = Environment(app)
