+++ nyaa/api_handler.py
@@ -6,7 +6,8 @@ import re
 
 import flask
 
-from nyaa import backend, bencode, db, forms, models, utils
+from nyaa import backend, bencode, forms, models, utils
+from nyaa.extensions import db
 from nyaa.views.torrents import _create_upload_category_choices
 
 api_blueprint = flask.Blueprint('api', __name__)
