+++ nyaa/models.py
@@ -15,7 +15,8 @@ from sqlalchemy.ext import declarative
 from sqlalchemy_fulltext import FullText
 from sqlalchemy_utils import ChoiceType, EmailType, PasswordType
 
-from nyaa import app, db
+from nyaa import app
+from nyaa.extensions import db
 from nyaa.torrents import create_magnet
 
 if app.config['USE_MYSQL']:
